extracted data parsing functions to /app/utils/data-parser.ts
